split of SquareBadges into SquareBadgeRow and SquareBadge
