+++ packages/gradient-block/src/components/index.ts
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './CssValueDisplay';
-export * from './SquareBadges';
+export * from './SquareBadgesRow';
 export * from './ColorPicker';
 export * from './AddColorButton';
 export * from './EditAndDeleteColorBox';
